target: Use size_t instead of unsigned in syscall interface.

+++ src/gdb_hostio.h
@@ -26,7 +26,7 @@ int hostio_reply(struct target_controller *tc, char *packet, int len);
 
 /* Interface to host system calls */
 int hostio_open(struct target_controller *,
-	        target_addr path, unsigned path_len,
+	        target_addr path, size_t path_len,
                 enum target_open_flags flags, mode_t mode);
 int hostio_close(struct target_controller *, int fd);
 int hostio_read(struct target_controller *,
@@ -36,18 +36,18 @@ int hostio_write(struct target_controller *,
 long hostio_lseek(struct target_controller *,
 	           int fd, long offset, enum target_seek_flag flag);
 int hostio_rename(struct target_controller *,
-	           target_addr oldpath, unsigned old_len,
-	           target_addr newpath, unsigned new_len);
+	           target_addr oldpath, size_t old_len,
+	           target_addr newpath, size_t new_len);
 int hostio_unlink(struct target_controller *,
-	           target_addr path, unsigned path_len);
+	           target_addr path, size_t path_len);
 int hostio_stat(struct target_controller *,
-	         target_addr path, unsigned path_len, target_addr buf);
+	         target_addr path, size_t path_len, target_addr buf);
 int hostio_fstat(struct target_controller *, int fd, target_addr buf);
 int hostio_gettimeofday(struct target_controller *,
 		         target_addr tv, target_addr tz);
 int hostio_isatty(struct target_controller *, int fd);
 int hostio_system(struct target_controller *,
-	           target_addr cmd, unsigned cmd_len);
+	           target_addr cmd, size_t cmd_len);
 
 #endif
